+++ simulation/src/wifi_helper/network.py
@@ -1,73 +1,88 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
 
+"""
+Fake Micropython network class
+
+See https://docs.micropython.org/en/latest/library/network.html
+"""
+
 import netifaces
 import random
 import re
 import socket
 import subprocess
-from sys import platform
-from typing import List, Optional, Tuple, Union
+import sys
+from typing import Any, List, Optional, Tuple, Union
 
 STA_IF = 1
 AP_IF = 2
 
+AUTH_OPEN = 1
+AUTH_WEP = 2
+AUTH_WPA_PSK = 3
+AUTH_WPA2_PSK = 4
+AUTH_WPA_WPA2_PSK = 5
+
 
 class NetworkHelper(object):
-    """docstring for Common"""
+    """docstring for NetworkHelper"""
     def __init__(self):
         pass
 
     @staticmethod
-    def _internal_scan() -> list:
-        networks = list()
+    def _internal_scan() -> List[Tuple[Union[str, int]]]:
+        """
+        Wrapper around internal functions depending on the platform type
 
-        if platform == "linux" or platform == "linux2":
-            networks = NetworkHelper._scan_linux()
-        elif platform == "darwin":
-            networks = NetworkHelper._scan_mac()
-        elif platform == "win32":
-            networks = NetworkHelper._scan_windows()
+        :returns:   List of found networks as tuple
+        :rtype:     List[Tuple[Union[str, int]]]
+        """
+        scan_result = list(dict())
+
+        if sys.platform == "linux" or sys.platform == "linux2":
+            scan_result = NetworkHelper._scan_linux()
+        elif sys.platform == "darwin":
+            scan_result = NetworkHelper._scan_mac()
+        elif sys.platform == "win32":
+            scan_result = NetworkHelper._scan_windows()
+        else:
+            scan_result = NetworkHelper._dummy_data()
+
+        networks = NetworkHelper._convert_scan_to_upy_format(data=scan_result)
 
         return networks
 
     @staticmethod
-    def _convert_dummy_data_to_raw(data: List[dict]) -> List[list]:
-        raw_nets = list()
+    def _scan_linux() -> List[dict]:
+        """
+        Perform WiFi scan on Linux systems.
 
-        for net in data:
-            this_raw = [
-                net['ssid'],
-                net['bssid'],
-                net['channel'],
-                net['RSSI'],
-                net['authmode'],
-                net['hidden']
-            ]
-            # [
-            #     'FRITZ!Box 7490',
-            #     '3810d517eb39',
-            #     11,
-            #     -random.randint(0, 100),
-            #     'WPA2-PSK',
-            #     False
-            # ]
+        This is currently not yet implemented, dummy data will be returned
 
-            raw_nets.append(this_raw)
-
-        return raw_nets
+        :returns:   List of found networks as dict
+        :rtype:     List[dict]
+        """
+        return NetworkHelper._dummy_data()
 
     @staticmethod
     def _scan_mac() -> List[dict]:
+        """
+        Perform WiFi scan on Mac systems.
+
+        Use command line interface of airport
+
+        :returns:   List of found networks as dict
+        :rtype:     List[dict]
+        """
         scan_result = subprocess.check_output(['/System/Library/PrivateFrameworks/Apple80211.framework/Versions/Current/Resources/airport', '--scan'])
 
         if len(scan_result) == 0:
             print('### Check WiFi to be active ###')
             print('Returning dummy data')
-            return NetworkHelper._dummy_data_raw()
+            return NetworkHelper._dummy_data()
 
         scan_result = [ele.decode('ascii') for ele in scan_result.splitlines()]
-        # print('scan_result: {}'.format(scan_result))
         # [
         #     b'                            SSID BSSID             RSSI CHANNEL HT CC SECURITY (auth/unicast/group)',
         #     b'                  TP-LINK_FBFC3C a0:f3:c1:fb:fc:3c -57  1,+1    Y  -- WPA(PSK/AES,TKIP/TKIP) WPA2(PSK/AES,TKIP/TKIP) ',
@@ -75,7 +90,6 @@ class NetworkHelper(object):
         # ]
 
         description = re.sub(' +', ' ', scan_result[0]).lstrip().split(' ')
-        # print('description: {}'.format(description))
         # [
         #   'SSID', 'BSSID', 'RSSI', 'CHANNEL', 'HT', 'CC', 'SECURITY',
         #   '(auth/unicast/group)'
@@ -84,7 +98,6 @@ class NetworkHelper(object):
         indexes = dict()
         for kw in description[0:-1]:
             indexes[kw.lower()] = scan_result[0].index(kw)
-        # print('indexes: {}'.format(indexes))
         # {
         #     'ssid': 28,
         #     'bssid': 33,
@@ -122,10 +135,7 @@ class NetworkHelper(object):
 
             info['hidden'] = True if info['ht'] == "Y" else False
 
-            # info['quality'] = dbm_to_quality(dBm=info['RSSI'])
-
             nets.append(info)
-            # print('info: {}'.format(info))
             # {
             #     'ssid': 'FRITZ!Box 7490',
             #     'bssid': '38:10:d5:17:eb:39',
@@ -134,11 +144,9 @@ class NetworkHelper(object):
             #     'ht': 'Y',
             #     'cc': 'DE',
             #     'authmode': 'WPA2(PSK/AES/AES)',
-            #     'hidden': True,
-            #     # 'quality': 86
+            #     'hidden': True
             # }
 
-        # print('nets: {}'.format(nets))
         # [
         #     {
         #         'ssid': 'FRITZ!Box 7490',
@@ -159,24 +167,40 @@ class NetworkHelper(object):
         #         'ht': 'Y',
         #         'cc': '--',
         #         'authmode': 'WPA(PSK/AES,TKIP/TKIP) WPA2(PSK/AES,TKIP/TKIP)',
-        #         'hidden': True,
-        #         # 'quality': 86
+        #         'hidden': True
         #     }
         # ]
 
-        return NetworkHelper._convert_dummy_data_to_raw(data=nets)
+        return nets
 
     @staticmethod
     def _scan_windows() -> List[dict]:
+        """
+        Perform WiFi scan on Windows systems.
+
+        Use command line interface of netsh.
+
+        This function might fail on non english or german language systems
+
+        :returns:   List of found networks as dict
+        :rtype:     List[dict]
+        """
         try:
             scan_result = subprocess.check_output(['netsh', 'wlan', 'show', 'networks', 'mode=bssid'])
         except Exception as e:
+            print('Failed to scan due to this error: {}'.format(e))
             print('### Check WiFi to be active ###')
             print('Returning dummy data')
-            return NetworkHelper._dummy_data_raw()
+            return NetworkHelper._dummy_data()
 
-        scan_result = [ele.decode('cp850').lstrip() for ele in scan_result.splitlines()]
-        # print(scan_result)
+        # netsh call returns report in local language, try to decode it
+        try:
+            scan_result = [ele.decode('cp850').lstrip() for ele in scan_result.splitlines()]
+        except Exception as e:
+            print('Failed to decode scan data due to this error: {}'.format(e))
+            print('Returning dummy data')
+            return NetworkHelper._dummy_data()
+
         # [
         #     '',
         #     'Schnittstellenname : WLAN ', 'Momentan sind 2 Netzwerke sichtbar.',
@@ -228,6 +252,7 @@ class NetworkHelper(object):
                 this_info = ele.split(':')
                 key = this_info[0].lstrip().rstrip().lower()
                 if len(this_info) > 2:
+                    # bssid keyword
                     val = ':'.join(this_info[1:]).lstrip().rstrip()
                 else:
                     val = this_info[1].lstrip().rstrip()
@@ -248,39 +273,42 @@ class NetworkHelper(object):
             quality = int(net['signal'][:-1])
             info['RSSI'] = NetworkHelper._quality_to_dbm(quality=quality)
 
-            info['channel'] = int(net['kanal'])
-
-            info['ht'] = '???'
-
-            info['cc'] = '???'
+            if 'channel' in net:
+                # english system language
+                info['channel'] = int(net['channel'])
+            elif 'kanal' in net:
+                # german system language
+                info['channel'] = int(net['kanal'])
+            else:
+                # other system language
+                info['channel'] = 1
 
             auth_key = [auth for auth in net.keys() if auth.startswith('auth')][0]
-            info['authmode'] = net[auth_key]
+            if len(auth_key):
+                # german or english system language
+                info['authmode'] = net[auth_key]
+            else:
+                # other system language
+                info['authmode'] = 'WPA2-Dummy'
 
             info['hidden'] = False
 
             nets.append(info)
-            # print('info: {}'.format(info))
             # {
             #     'ssid': 'FRITZ!Box 7490',
             #     'bssid': '38:10:d5:17:eb:39',
             #     'RSSI': '87%',
             #     'channel': '11',
-            #     'ht': '???',
-            #     'cc': 'WAP',
             #     'authmode': 'WPA2-Personal',
             #     'hidden': False
             # }
 
-        # print('nets: {}'.format(nets))
         # [
         #     {
         #         'ssid': 'FRITZ!Box 7490',
         #         'bssid': '38:10:d5:17:eb:39',
         #         'RSSI': '85%',
         #         'channel': '11',
-        #         'ht': '???',
-        #         'cc': '???',
         #         'authmode': 'WPA2-Personal',
         #         'hidden': False
         #     },
@@ -289,14 +317,12 @@ class NetworkHelper(object):
         #         'bssid': 'a0:f3:c1:fb:fc:3c',
         #         'RSSI': '83%',
         #         'channel': '1',
-        #         'ht': '???',
-        #         'cc': '???',
         #         'authmode': 'WPA2-Personal',
         #         'hidden': False
         #     }
         # ]
 
-        return NetworkHelper._convert_dummy_data_to_raw(data=nets)
+        return nets
 
     @staticmethod
     def _quality_to_dbm(quality: int) -> int:
@@ -310,33 +336,24 @@ class NetworkHelper(object):
         :rtype:     int
         """
         # https://stackoverflow.com/questions/15797920/how-to-convert-wifi-signal-strength-from-quality-percent-to-rssi-dbm/30585711
-        """
-        dBm = 0
-        if quality <= 0:
-            # very bad signal
-            dBm = -100
-        elif quality >= 100:
-            dBm = -50
-        else:
-            dBm = (quality / 2) - 100
-
-        return dBm
-        """
         return min(max((quality / 2) - 100, -100), -50)
 
     @staticmethod
-    def _scan_linux() -> List[dict]:
-        return NetworkHelper._dummy_data_raw()
-
-    @staticmethod
     def _dummy_data() -> List[dict]:
+        """
+        Get dummy WiFi scan data.
+
+        RSSI value is fluctuating between 0 and 100
+
+        :returns:   List of found networks as dict
+        :rtype:     List[dict]
+        """
         nets = [
             {
                 'ssid': 'TP-LINK_FBFC3C',
                 'RSSI': -random.randint(0, 100),
                 'bssid': 'a0f3c1fbfc3c',
                 'authmode': 'WPA/WPA2-PSK',
-                # 'quality': random.randint(0, 100),
                 'channel': 1,
                 'hidden': False
             },
@@ -345,7 +362,6 @@ class NetworkHelper(object):
                 'RSSI': -random.randint(0, 100),
                 'bssid': '3810d517eb39',
                 'authmode': 'WPA2-PSK',
-                # 'quality': random.randint(0, 100),
                 'channel': 11,
                 'hidden': False
             }
@@ -354,33 +370,61 @@ class NetworkHelper(object):
         return nets
 
     @staticmethod
-    def _dummy_data_raw() -> List[list]:
+    def _dummy_data_upy() -> List[list]:
+        """
+        Get dummy WiFi scan data in same style as Mircopython would return it
+
+        :returns:   List of lists with scan data as reported on Micropython
+        :rtype:     List[Union[str, int]]
+        """
         dummy_data = NetworkHelper._dummy_data()
-        nets_raw = NetworkHelper._convert_dummy_data_to_raw(data=dummy_data)
-        """
-        [
-            [
-                'TP-LINK_FBFC3C',
-                'a0f3c1fbfc3c',
-                1,
-                -random.randint(0, 100),
-                'WPA/WPA2-PSK',
-                False
-            ],
-            [
-                'FRITZ!Box 7490',
-                '3810d517eb39',
-                11,
-                -random.randint(0, 100),
-                'WPA2-PSK',
-                False
-            ]
-        ]
-        """
+        nets_raw = NetworkHelper._convert_scan_to_upy_format(data=dummy_data)
+
         return nets_raw
 
     @staticmethod
+    def _convert_scan_to_upy_format(data: List[dict]) -> List[Tuple[Union[str, int]]]:
+        """
+        Convert dummy data to raw data as Mircopython scan would return
+
+        :param      data:  List of complete WiFi scan data
+        :type       data:  List[dict]
+
+        :returns:   List of tuples with scan data as reported on Micropython
+        :rtype:     List[Tuple[Union[str, int]]]
+        """
+        raw_nets = list()
+
+        for net in data:
+            this_raw = (
+                net['ssid'],
+                net['bssid'],
+                net['channel'],
+                net['RSSI'],
+                net['authmode'],
+                net['hidden']
+            )
+            # (
+            #     'FRITZ!Box 7490',
+            #     '3810d517eb39',
+            #     11,
+            #     -random.randint(0, 100),
+            #     'WPA2-PSK',
+            #     False
+            # )
+
+            raw_nets.append(this_raw)
+
+        return raw_nets
+
+    @staticmethod
     def _gather_ifconfig_data() -> Tuple[str, str, str, str]:
+        """
+        Gather ifconfig data
+
+        :returns:   fifconfig data in same style as Micropython would return it
+        :rtype:     Tuple[str, str, str, str]
+        """
         ip = '192.168.0.1'
         subnet = '255.255.255.0'
         gateway = '192.168.0.1'
@@ -417,6 +461,44 @@ class Station(NetworkHelper):
     def __init__(self):
         self._connected = False
         self._active = False
+        self._connected_network = None
+
+    def active(self, is_active: Optional[bool] = None) -> Union[None, bool]:
+        """
+        Get/set WiFi status
+
+        :param      is_active:  Flag to change state of WiFi
+        :type       is_active:  Optional[bool]
+
+        :returns:   State of WiFi interface
+        :rtype:     Union[None, bool]
+        """
+        if is_active is not None:
+            self._active = is_active
+        else:
+            return self._active
+
+    def connect(self,
+                service_id: str,
+                key: Optional[str] = None,
+                **kwargs) -> None:
+        """
+        Connect to a network
+
+        :param      service_id:  The service identifier (BSSID/MAC address)
+        :type       service_id:  str
+        :param      key:         The key (password)
+        :type       key:         Optional[str]
+        :param      kwargs:      The keywords arguments
+        :type       kwargs:      dictionary
+        """
+        # remember connected BSSID
+        self._connected_network = service_id
+        self.connected = True
+
+    def disconnect(self) -> None:
+        """Disconnect from network."""
+        self.connected = False
 
     def isconnected(self) -> bool:
         """
@@ -426,6 +508,61 @@ class Station(NetworkHelper):
         :rtype:     bool
         """
         return self.connected
+
+    def scan(self) -> List[Tuple[Union[str, int]]]:
+        """
+        Return WiFi scan data
+
+        :returns:   List of tuples with discovered service parameters
+        :rtype:     List[Tuple[Union[str, int]]]
+        """
+        return self._internal_scan()
+
+    def status(self, param: Optional[str] = None) -> Union[str, int, tuple]:
+        """
+        Query dynamic status information of the interface
+
+        :param      param:  The status parameter to retrieve
+        :type       param:  Optional[str]
+
+        :returns:   Status information
+        :rtype:     Union[str, int, tuple]
+        """
+        if param is not None:
+            return self.isconnected()
+        else:
+            pass
+
+    def ifconfig(self,
+                 value: Optional[Tuple[str, str, str, str]] = None) -> Union[None, Tuple[str, str, str, str]]:
+        """
+        Get/set IP-level network interface parameters
+
+        :param      value:  The value to set
+        :type       value:  Optional[Tuple[str, str, str, str]]
+
+        :returns:   A 4-tuple with information about IO, subnet, gateway, DNS
+        :rtype:     Union[None, Tuple[str, str, str, str]]
+        """
+        if value is not None:
+            print('Change of ifconfig not supported')
+        else:
+            return self._gather_ifconfig_data()
+
+    def config(self, param: str) -> Union[str, int]:
+        """
+        Get/set general network interface parameters
+
+        :param      param:  The status parameter to set
+        :type       param:  Optional[str]
+
+        :returns:   Status information
+        :rtype:     Union[str, int, tuple]
+        """
+        if param == 'essid':
+            return self._connected_network
+        else:
+            pass
 
     @property
     def connected(self) -> bool:
@@ -447,41 +584,89 @@ class Station(NetworkHelper):
         """
         self._connected = value
 
-    def active(self, value=None) -> Union[None, bool]:
-        if value is not None:
-            print('Change working mode to: {}'.format(value))
-            self._active = value
-        else:
-            return self._active
-
-    def scan(self) -> list:
-        networks = list()
-
-        networks = self._internal_scan()
-
-        return networks
-
-    def ifconfig(self, value: Optional[Tuple[str, str, str, str]] = None) -> Union[None, Tuple[str, str, str, str]]:
-        if value is not None:
-            print('Change of ifconfig not supported')
-        else:
-            return self._gather_ifconfig_data()
-
 
 class Client(NetworkHelper):
     """docstring for Client"""
-    def __init__(self):
-        pass
+    valid_kwargs = ['_essid', '_authmode', '_password', '_channel', '_active']
 
-    def ifconfig(self, value: Optional[Tuple[str, str, str, str]] = None) -> Union[None, Tuple[str, str, str, str]]:
+    def __init__(self):
+        # for key in Client.valid_kwargs:
+        #     setattr(self, key, None)
+
+        self._essid = None
+        self._authmode = None
+        self._password = None
+        self._channel = None
+        self._active = None
+        self._active = False
+
+    def active(self, is_active: Optional[bool] = None) -> Union[None, bool]:
+        """
+        Get/set WiFi status
+
+        :param      is_active:  Flag to change state of WiFi
+        :type       is_active:  Optional[bool]
+
+        :returns:   State of WiFi interface
+        :rtype:     Union[None, bool]
+        """
+        if is_active is not None:
+            self._active = is_active
+        else:
+            return self._active
+
+    def ifconfig(self,
+                 value: Optional[Tuple[str, str, str, str]] = None) -> Union[None, Tuple[str, str, str, str]]:
+        """
+        Get/set IP-level network interface parameters
+
+        :param      value:  The value to set
+        :type       value:  Optional[Tuple[str, str, str, str]]
+
+        :returns:   A 4-tuple with information about IO, subnet, gateway, DNS
+        :rtype:     Union[None, Tuple[str, str, str, str]]
+        """
         if value is not None:
             print('Change of ifconfig not supported')
         else:
             return self._gather_ifconfig_data()
 
+    def config(self, **kwargs) -> None:
+        print('Config Accesspoint with: {}'.format(kwargs))
+        for key, val in kwargs.items():
+            key_name = '_{}'.format(key)
+            if key_name not in Client.valid_kwargs:
+                raise TypeError("Invalid keyword argument {}".format(key))
+            print('Set {} to {}'.format(key_name, val))
+            setattr(self, key_name, val)
 
-class WLAN(Client, Station):
-    """docstring for WLAN"""
+
+class WLAN(object):
+    """
+    WLAN would inherit from Station and Client.
+
+    Due to the diamond problem this hack fakes an inheritance of another object
+    by trying to find the attribute in its child classes in case it has not
+    been found in the parent class. The lookup is done in that class, which has
+    been set as "true" child during the init.
+
+    See https://en.wikipedia.org/wiki/Multiple_inheritance#The_diamond_problem
+    """
     def __init__(self, val: int) -> None:
-        Station.__init__(self)
-        Client.__init__(self)
+        self._type = None
+        self._type_name = None
+
+        if val == STA_IF:
+            self._type = Station()
+            self._type_name = STA_IF
+        elif val == AP_IF:
+            self._type = Client()
+            self._type_name = AP_IF
+
+    def __getattr__(self, item: Any) -> Any:
+        if self._type_name == STA_IF:
+            return super(Station, self._type).__getattribute__(item)
+        elif self._type_name == AP_IF:
+            return super(Client, self._type).__getattribute__(item)
+        else:
+            raise AttributeError("WLAN has no attribute '{}'".format(item))